Split stack methods and clean linked groups
